# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AIVEC Conference API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [all, upcoming, past, open]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{slug}:
    get:
      operationId: getEvent
      tags: [events]
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventWithFields"
  /events/id/{id}:
    patch:
      operationId: updateEvent
      tags: [events]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /events/id/{id}/fields:
    put:
      operationId: setEventFields
      tags: [events]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventFieldsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventField"
  /events/id/{id}/register:
    post:
      operationId: registerForEvent
      tags: [registrations]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
  /events/id/{id}/registrations:
    get:
      operationId: listEventRegistrations
      tags: [registrations]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
...
  /registrations/{id}:
    patch:
      operationId: updateRegistration
      tags: [registrations]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
...
  /blog:
    get:
      operationId: listBlog
      tags: [blog]
      parameters:
        - in: query
          name: published
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlog
      tags: [blog]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/{slug}:
    get:
      operationId: getBlog
      tags: [blog]
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
  /blog/id/{id}:
    patch:
      operationId: updateBlog
      tags: [blog]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
    delete:
      operationId: deleteBlog
      tags: [blog]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
  /contact-submissions:
    get:
      operationId: listContactSubmissions
      tags: [contact]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactSubmission"
    post:
      operationId: createContactSubmission
      tags: [contact]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactSubmissionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmission"
  /contact-submissions/{id}:
    delete:
      operationId: deleteContactSubmission
      tags: [contact]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
...
  schemas:
...
    Event:
      type: object
      required: [id, slug, status]
      properties:
        id: { type: integer }
        slug: { type: string }
        titleEn: { type: string }
        titleAr: { type: string }
        summaryEn: { type: string }
        summaryAr: { type: string }
        bodyEn: { type: string }
        bodyAr: { type: string }
        venueEn: { type: string }
        venueAr: { type: string }
        startsAt: { type: ["string", "null"], format: date-time }
        endsAt: { type: ["string", "null"], format: date-time }
        status: { type: string, enum: [draft, coming_soon, open, closed, past] }
        featured: { type: boolean }
        posterUrl: { type: ["string", "null"] }
        seoTitleEn: { type: ["string", "null"] }
        seoTitleAr: { type: ["string", "null"] }
        seoDescEn: { type: ["string", "null"] }
        seoDescAr: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    EventField:
      type: object
      required: [id, eventId, fieldKey, fieldType, order]
      properties:
        id: { type: integer }
        eventId: { type: integer }
        fieldKey: { type: string }
        fieldType:
          type: string
          enum: [short_text, long_text, email, phone, number, dropdown, radio, checkbox, date]
        labelEn: { type: string }
        labelAr: { type: string }
        helpEn: { type: string }
        helpAr: { type: string }
        placeholderEn: { type: string }
        placeholderAr: { type: string }
        required: { type: boolean }
        order: { type: integer }
        options: { type: array, items: { $ref: "#/components/schemas/EventFieldOption" } }
    EventFieldOption:
      type: object
      required: [value, labelEn, labelAr]
      properties:
        value: { type: string }
        labelEn: { type: string }
        labelAr: { type: string }
    EventWithFields:
      type: object
      required: [event, fields]
      properties:
        event: { $ref: "#/components/schemas/Event" }
        fields:
          type: array
          items:
            $ref: "#/components/schemas/EventField"
    EventInput:
      type: object
      required: [slug, titleEn, titleAr]
      properties:
        slug: { type: string }
        titleEn: { type: string }
        titleAr: { type: string }
        summaryEn: { type: string }
        summaryAr: { type: string }
        bodyEn: { type: string }
        bodyAr: { type: string }
        venueEn: { type: string }
        venueAr: { type: string }
        startsAt: { type: ["string", "null"], format: date-time }
        endsAt: { type: ["string", "null"], format: date-time }
        status: { type: string, enum: [draft, coming_soon, open, closed, past] }
        featured: { type: boolean }
        posterUrl: { type: ["string", "null"] }
        seoTitleEn: { type: ["string", "null"] }
        seoTitleAr: { type: ["string", "null"] }
        seoDescEn: { type: ["string", "null"] }
        seoDescAr: { type: ["string", "null"] }
    EventUpdate:
      type: object
      properties:
        slug: { type: string }
        titleEn: { type: string }
        titleAr: { type: string }
        summaryEn: { type: string }
        summaryAr: { type: string }
        bodyEn: { type: string }
        bodyAr: { type: string }
        venueEn: { type: string }
        venueAr: { type: string }
        startsAt: { type: ["string", "null"], format: date-time }
        endsAt: { type: ["string", "null"], format: date-time }
        status: { type: string, enum: [draft, coming_soon, open, closed, past] }
        featured: { type: boolean }
        posterUrl: { type: ["string", "null"] }
        seoTitleEn: { type: ["string", "null"] }
        seoTitleAr: { type: ["string", "null"] }
        seoDescEn: { type: ["string", "null"] }
        seoDescAr: { type: ["string", "null"] }
    EventFieldInput:
      type: object
      required: [fieldKey, fieldType, labelEn, labelAr]
      properties:
        fieldKey: { type: string }
        fieldType:
          type: string
          enum: [short_text, long_text, email, phone, number, dropdown, radio, checkbox, date]
        labelEn: { type: string }
        labelAr: { type: string }
        helpEn: { type: string }
        helpAr: { type: string }
        placeholderEn: { type: string }
        placeholderAr: { type: string }
        required: { type: boolean }
        order: { type: integer }
        options: { type: array, items: { $ref: "#/components/schemas/EventFieldOption" } }
    EventFieldsInput:
      type: object
      required: [fields]
      properties:
        fields:
          type: array
          items:
            $ref: "#/components/schemas/EventFieldInput"
    Registration:
      type: object
      required: [id, eventId, userId, status, createdAt, answers]
      properties:
        id: { type: integer }
        eventId: { type: integer }
        userId: { type: integer }
        userEmail: { type: ["string", "null"] }
        userName: { type: ["string", "null"] }
        eventTitleEn: { type: ["string", "null"] }
        eventTitleAr: { type: ["string", "null"] }
        status: { type: string, enum: [submitted, reviewed, accepted, rejected] }
        notes: { type: ["string", "null"] }
        answers: {}
        createdAt: { type: string, format: date-time }
    RegistrationInput:
      type: object
      required: [answers]
      properties:
        answers: {}
    RegistrationUpdate:
      type: object
      properties:
        status: { type: string, enum: [submitted, reviewed, accepted, rejected] }
        notes: { type: ["string", "null"] }
...
    BlogPost:
      type: object
      required: [id, slug, titleEn, titleAr, published]
      properties:
        id: { type: integer }
        slug: { type: string }
        titleEn: { type: string }
        titleAr: { type: string }
        excerptEn: { type: string }
        excerptAr: { type: string }
        bodyEn: { type: string }
        bodyAr: { type: string }
        tags: { type: array, items: { type: string } }
        coverUrl: { type: ["string", "null"] }
        authorName: { type: string }
        published: { type: boolean }
        publishedAt: { type: ["string", "null"], format: date-time }
        seoTitleEn: { type: ["string", "null"] }
        seoTitleAr: { type: ["string", "null"] }
        seoDescEn: { type: ["string", "null"] }
        seoDescAr: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    BlogInput:
      type: object
      required: [slug, titleEn, titleAr]
      properties:
        slug: { type: string }
        titleEn: { type: string }
        titleAr: { type: string }
        excerptEn: { type: string }
        excerptAr: { type: string }
        bodyEn: { type: string }
        bodyAr: { type: string }
        tags: { type: array, items: { type: string } }
        coverUrl: { type: ["string", "null"] }
        authorName: { type: string }
        published: { type: boolean }
        publishedAt: { type: ["string", "null"], format: date-time }
        seoTitleEn: { type: ["string", "null"] }
        seoTitleAr: { type: ["string", "null"] }
        seoDescEn: { type: ["string", "null"] }
        seoDescAr: { type: ["string", "null"] }
    BlogUpdate:
      type: object
      properties:
        slug: { type: string }
        titleEn: { type: string }
        titleAr: { type: string }
        excerptEn: { type: string }
        excerptAr: { type: string }
        bodyEn: { type: string }
        bodyAr: { type: string }
        tags: { type: array, items: { type: string } }
        coverUrl: { type: ["string", "null"] }
        authorName: { type: string }
        published: { type: boolean }
        publishedAt: { type: ["string", "null"], format: date-time }
        seoTitleEn: { type: ["string", "null"] }
        seoTitleAr: { type: ["string", "null"] }
        seoDescEn: { type: ["string", "null"] }
        seoDescAr: { type: ["string", "null"] }
    ContactSubmission:
      type: object
      required: [id, name, email, message, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        subject: { type: string }
        message: { type: string }
        createdAt: { type: string, format: date-time }
    ContactSubmissionInput:
      type: object
      required: [name, email, message]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, minLength: 3 }
        phone: { type: string }
        subject: { type: string }
        message: { type: string, minLength: 1 }
...