# lib/api-zod/src/generated/types/blogUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIVEC Conference API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogUpdate {
  slug?: string;
  titleEn?: string;
  titleAr?: string;
  excerptEn?: string;
  excerptAr?: string;
  bodyEn?: string;
  bodyAr?: string;
  tags?: string[];
  /** @nullable */
  coverUrl?: string | null;
  authorName?: string;
  published?: boolean;
  /** @nullable */
  publishedAt?: Date | null;
  /** @nullable */
  seoTitleEn?: string | null;
  /** @nullable */
  seoTitleAr?: string | null;
  /** @nullable */
  seoDescEn?: string | null;
  /** @nullable */
  seoDescAr?: string | null;
}