# lib/api-zod/src/generated/types/contactSubmissionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIVEC Conference API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactSubmissionInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 3 */
  email: string;
  phone?: string;
  subject?: string;
  /** @minLength 1 */
  message: string;
}