# lib/api-zod/src/generated/types/eventFieldInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AIVEC Conference API
 * OpenAPI spec version: 0.1.0
 */
import type { EventFieldInputFieldType } from './eventFieldInputFieldType';
import type { EventFieldOption } from './eventFieldOption';

export interface EventFieldInput {
  fieldKey: string;
  fieldType: EventFieldInputFieldType;
  labelEn: string;
  labelAr: string;
  helpEn?: string;
  helpAr?: string;
  placeholderEn?: string;
  placeholderAr?: string;
  required?: boolean;
  order?: number;
  options?: EventFieldOption[];
}